bugzid: 2799 refactor uniprot load to work faster.
git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@26911 908b024a-c82b-635c-c293-f2b15b23277c

+++ lib/DB_functions/SPL/p_check_drop_go_root_term.sql
@@ -27,25 +27,31 @@ create procedure p_check_drop_go_root_term (
 	if (gotermZdbId not in ('ZDB-TERM-091209-6070','ZDB-TERM-091209-2432','ZDB-TERM-091209-4029')) then
 
 	    -- find out the root term's zdb id, which is in the same category
+	    
+	    let goRootTermZdbId = "ok";
+ 
 	    select u.term_zdb_id
 	      into goRootTermZdbId
 	      from term u, term n
 	     where n.term_zdb_id = gotermZdbId
                and u.term_ontology = n.term_ontology
 	       and u.term_ont_id in ("GO:0005575", "GO:0003674", "GO:0008150");
-
+	    
+	    if (gotermZdbId not like 'ok') then
+	    
 	
 	    -- delete the root term annotation
 	    delete from zdb_active_data 
-	          where zactvd_zdb_id in 
-		       	(select mrkrgoev_zdb_id 
-			   from marker_go_term_evidence
-			  where mrkrgoev_mrkr_zdb_id = mrkrZdbId
-			    and mrkrgoev_term_zdb_id = goRootTermZdbId
+	          where exists (select 'x' from marker_go_Term_evidence 
+		  	       	       where mrkrgoev_zdb_id = zactvd_zdb_id
+				       and mrkrgoev_mrkr_zdb_id = mrkrZdbId
+			    	       and mrkrgoev_term_zdb_id = goRootTermZdbId
 			);
 
 	      execute procedure p_drop_go_root_term_attribution (mrkrZdbId);
 	    
+	    end if;
+
 	end if; 
 
 end procedure;
